backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==4.4.2
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
moviepy==1.0.3
numpy==1.26.4
pillow==11.0.0
proglog==0.1.10
psycopg2-binary==2.9.10
pyarrow==18.0.0
pylance==0.19.1
requests==2.32.3
sqlparse==0.5.1
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
